Added a function to launcherUtils that returns the correct executable name corresponding to user's OS.
This function is now used by other classes to store the executable names in member variables. Relevant code now refers to the member variables instead of the previously hardcoded exe names.

In addition, I adjusted download links to point to executables (with the platform agnostic exe names) although that will require building the launcher for Linux on each release.