Moved pysimplegui layout code to its own utility class, layoutUtils. layoutUtils has one public method, generate(), along with several private helper methods.
My goals with this commit were to:
Separate layout code from the main code. This improves readability for both sets of code.
Break up the layout code into separate functions and variables so that it is easier to read and modify.
Assemble the layout elements so that they resemble the actual UI, as it is suggested in the pysimplegui docs. Read the returns for side_panel(), main_panel(), and main_panel()'s variable button_panel_with_mod_info for this idea in action.